Move TelegramUtils to separate module
